backend/requirements.txt
Flask
sqlalchemy
Flask-SQLAlchemy
Flask-Cors
Flask-Migrate
python-dotenv
gunicorn